# speedup DeserializerCache