@SamSaffron 2021-02-26
looks good, a bit surprising but makes sense. I guess this is technically a regression if for some reason one table has a column called `a,b` but we can just accept that these pathological cases are not handled