@ermolaev 2021-02-28
I agree

PostgreSQL not supported alias with contains `,` 
```sql
SELECT  name AS 'a,b' FROM "products"
PG::SyntaxError: ERROR:  syntax error at or near "'a,b'"
```

but support column names in tables

```SQL
CREATE TABLE tbl (
 "a,b" text
);
insert into tbl values ('t');
```

and this is hard to work in ruby because we don't may create a valid method with `,` and earlier mini-sql raises an exception on it, so there is no regression
```sql
MINI_SQL.query("select * from tbl ")
invalid attribute name `a,b'
mini_sql/postgres/deserializer_cache.rb:74:in `attr_accessor'
```